sdk: go: move engineconn into internal package
