+++ sdk/go/internal/engineconn/engineconn.go
